implement custom opId-to-function mapping
Swagger schemas enforce unique values for operationIds, which are
per-HTTP Verb, but in the case of a liberator resource function,
we have just one large function which handles the entire resource.

The new mapping function allows the operationIds to be unique by
appending arbitrary data to the end of them, after a *. Only the
text prior to the * denotes the namespace and function identifier.